requirements.txt
google-generativeai
grpcio
grpcio-tools
flask